requirements.txt
akshare==1.1.98
backtrader==1.9.76.123
beautifulsoup4==4.10.0
Bottleneck==1.3.2
certifi==2021.10.8
charset-normalizer==2.0.7
cloudpickle==2.0.0
colorama==0.4.4
cycler==0.11.0
decorator==5.1.0
et-xmlfile==1.1.0
graphviz==0.17
html5lib==1.1
idna==3.3
jsonpath==0.82
kiwisolver==1.3.2
lxml==4.6.4
matplotlib==3.2.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mplfinance==0.12.7a17
neat-python==0.92
numexpr==2.7.3
numpy==1.21.2
openpyxl==3.0.9
pandas==1.3.4
Pillow==8.4.0
pip==21.0.1
py-mini-racer==0.6.0
pyparsing==3.0.4
pypinyin==0.43.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
setuptools==58.0.4
six==1.16.0
soupsieve==2.2.1
tabulate==0.8.9
tqdm==4.62.3
urllib3==1.26.7
webencodings==0.5.1
websocket-client==1.2.1
wheel==0.37.0
wincertstore==0.2
xlrd==2.0.1
